% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/green_vehicles.R
\docType{data}
\name{green_vehicles}
\alias{green_vehicles}
\title{Green vehicle information from the EPA}
\format{
A tibble with 19 variables(columns) and 6178 observations(row):
\describe{
\item{year}{Model year}
\item{model}{Vehicle make and mode}
\item{displ}{– engine displacement in liters}
\item{cyl}{number of engine cylinders}
\item{trans}{transmission type plus number of gears}
\item{drive}{2-wheel Drive, 4-wheel drive/all-wheel drive}
\item{fuel}{Gasoline, Electricity, Gasolin/Electricity, Diesel, Ethanol/Gas, Hydrogen}
\item{cert_region}{Region of certification:
  \itemize{
    \item CA - California
    \item CE - California + NLEV (Northeast trading area)
    \item CF - Clean Fuel Vehicle
    \item CL - Calif. + NLEV(All states)
    \item FA - Federal All Altitude
    \item FC - Tier 2 Federal and Aclif.
    \item NF - CFV + NLEV(ASTR) + Calif
    \item NL - NLEV(All states)
}}
\item{stnd}{vehicle emissions standard code. See Stnd Description.}
\item{stnd_description}{vehicle emissions standard description. See: \url{https://www.epa.gov/greenvehicles/federal-and-california-light-duty-vehicle-emissions-standards-air-pollutants} }
\item{underhood_id}{engine family or test group ID. See: \url{http://www.fueleconomy.gov/feg/findacarhelp.shtml#airPollutionScore}}
\item{veh_class}{EPA vehicle class. See: \url{http://www.fueleconomy.gov/feg/findacarhelp.shtml#epaSizeClass}}
\item{air_pollution_score}{Smog Rating. See: \url{https://www.epa.gov/greenvehicles/smog-rating}}
\item{city_mpg}{city fuel economy in miles per gallon}
\item{hwy_mpg}{highway fuel economy in miles per gallon}
\item{cmb_mpg}{combined highway/city fuel economy in miles per gallon}
\item{greenhouse_gas_score}{Greenhouse Gas Rating. See: \url{https://www.epa.gov/greenvehicles/greenhouse-gas-rating}}
\item{smartway}{Yes, No, or Elite. See: \url{https://www.epa.gov/greenvehicles/consider-smartway-vehicle}}
\item{comb_co2}{combined city/highway CO2 tailpipe emissions in grams per mile}
}
}
\source{
\url{http://www.fueleconomy.gov/feg/download.shtml}

\url{https://www.fueleconomy.gov/feg/EPAGreenGuide/GreenVehicleGuideDocumentation.pdf}
}
\usage{
green_vehicles
}
\description{
This dataset contains vehicle information published by the EPA in their
their annual "EPA Green Vehicle Guide". The dataset currently holds information for the
2023, 2022, and 2021 vehicles.
}
\keyword{datasets}
